Added Tap BPM
There is now a BPM indicator at the top of the screen. Tapping on it 8 times will set the BPM to the tempo you tap at. After a couple of seconds the number of taps resets so you don't get into a situation where you set the tempo instantly to something really small.