Finalised post effects system.
There are now 10 post effects banks. You can access these banks via the Post FX tabs. Click and drag a post effect to one of the ten slots at the bottom of the screen and it'll be assigned to a slot. To de-allocate an effect you can click the button on the right hand side. You can select an effect by clicking the slot.

Next steps will include some demo post effects and the post effects template file. (for now see the example tox in Lib/base_addpost/